refactor(first-run): consolidate IdentifyStep tests into identify-step.component.test.tsx
